cleanup AzCopyCore sample
