+++ samples/AzCopyCore/AzCopyCore/StorageClient.cs
@@ -1,4 +1,8 @@
-﻿using System.Azure.Authentication;
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Azure.Authentication;
 using System.Azure.Storage.Requests;
 using System.Buffers.Cryptography;
 using System.Buffers.Text;
