+++ samples/AzCopyCore/AzCopyCore/StorageRequests.cs
@@ -1,5 +1,8 @@
-﻿using System.Azure.Authentication;
-using System.Buffers;
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Azure.Authentication;
 using System.Buffers.Text;
 using System.Buffers.Transformations;
 using System.IO;
